written server chatroom testcode in main.py
