+++ Protocol/server_chatroom.py
@@ -1,37 +1,39 @@
-import socket, threading, select
+import select, socket, threading
 
-class ServerChatroom:
-    def __init__(self, host: str, port: int) -> None:
-        self.sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        self.sock.bind((host, port))
-        self.sock.listen()
-        self.conn, addr = self.sock.accept()
-        self.message_log: list[bytes] = []
-        print(f"Connected by {addr}")
-        self.is_alive = True
+messages = [ ] # outgoing message queue
+is_running = False
+sock: socket.socket = None
+select_inout = [ ]
 
-        def receive():
-            while self.is_alive:
-                data = self.conn.recv(1024)
+def start_chatroom(ip: str, port: int):
+    global sock, is_running
+    sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    sock.setblocking(False)
+    sock.bind((ip, port))
+    sock.listen()
+    select_inout.append(sock)
+    is_running = True
+    chatroom = threading.Thread(target=chatroom_thread)
+    chatroom.start()
+
+def chatroom_thread():
+    while is_running:
+        readable, writable, _ = select.select(select_inout, select_inout, [])
+        for s in readable:
+            if s is sock:
+                conn, addr = s.accept()
+                print(f"connected from {addr}")
+                select_inout.append(conn)
+            else: 
+                data = s.recv(1024)
+                if data:
+                    print(data)
                 if not data:
-                    print("THIS CHAT HAS BEEN TERMINATED")
-                    self.conn.close()
-                    self.is_alive = False
-                    break
-                print(data)
-                self.message_log.append(data)
-
-        self.receive_thread = threading.Thread(target=receive)
-        self.receive_thread.start()
-
-    def terminate(self):
-        self.conn.close()
-        self.sock.shutdown(socket.SHUT_RDWR)
-        self.sock.close()
-        self.is_alive = False
-
-    def send(self, message):
-        self.message_log.append(message)
-        self.conn.send(message)
-
-
+                    print("THIS CHATROOM IS TERMINATED")
+                    sock.close()
+                    return
+        for s in writable:
+            if messages:
+                s.sendall(messages[0]) 
+                messages.pop(0)
+    sock.close()
