+++ main.py
@@ -1,21 +1,14 @@
-from Protocol import ServerChatroom, ClientChatroom
+from Protocol import server_chatroom, ClientChatroom
 
 chatroom_IP = "127.0.0.1"
 chatroom_PORT = 5000
 
-#s = ServerChatroom(chatroom_IP, chatroom_PORT)
-s = ClientChatroom(chatroom_IP, chatroom_PORT)
+server_chatroom.start_chatroom(chatroom_IP, chatroom_PORT)
 
 while True:
     m = input()
-
-    if not s.is_alive: break
+    if not server_chatroom.is_running: break
     if m == "END": 
-        s.terminate()
+        server_chatroom.is_running = False
         break
-
-    if m == "LOG":
-        for i in s.message_log: print(i)
-        continue
-
-    s.send(m.encode())
+    server_chatroom.messages.append(m.encode())
